requirements.txt
pandas
streamlit
plotly
tqdm
azure-ai-inference
azure-core
requests
python-dotenv